refactor: move modal components into the ui/modals subdirectory
