+++ src/fileMenu.ts
@@ -3,7 +3,7 @@ import { deleteXoppAndPdf, findCorrespondingXoppToPdf, openXournalppFile, rename
 import CreateXoppModalManager from "src/ui/managers/create-xopp-modal-manager";
 import XoppPlugin from "src/main";
 import { exportXoppToPDF } from "./xopp2pdf";
-import RenameModal from "./modals/RenameModal";
+import RenameModal from "./ui/modals/rename-modal";
 
 export function addXournalppOptionsToFileMenu(menu: Menu, file: TFile | TFolder, plugin: XoppPlugin) {
     if (file instanceof TFile) {
