+++ src/ui/modals/new-file-placholder-help-modal.ts
