+++ src/ui/modals/search-xopp-modal.ts
