+++ src/ui/modals/template-editing-modal.ts
@@ -1,7 +1,7 @@
 import XoppPlugin from "src/main";
 import { App, ButtonComponent, Modal, Setting } from "obsidian";
 import { parseTemplateFile, fetchTemplates } from "src/ui/managers/template-editing-modal-manager";
-import { TemplateSpec } from "./TemplateCreationModal";
+import { TemplateSpec } from "./template-creation-modal";
 import { createTemplate } from "src/ui/managers/template-creation-modal-manager";
 
 export default class TemplateEditingModal extends Modal {
